refactor(PlotterBase): separate file
